+++ test/terra/backends/test_qasm_stabilizer_simulator.py
@@ -32,6 +32,7 @@ from test.terra.backends.qasm_simulator.qasm_noise import QasmPauliNoiseTests
 from test.terra.backends.qasm_simulator.qasm_noise import QasmResetNoiseTests
 # Snapshot tests
 from test.terra.backends.qasm_simulator.qasm_snapshot import QasmSnapshotStatevectorTests
+from test.terra.backends.qasm_simulator.qasm_snapshot import QasmSnapshotDensityMatrixTests
 from test.terra.backends.qasm_simulator.qasm_snapshot import QasmSnapshotStabilizerTests
 # Other tests
 from test.terra.backends.qasm_simulator.qasm_method import QasmMethodTests
@@ -49,6 +50,7 @@ class TestQasmStabilizerSimulator(common.QiskitAerTestCase,
                                   QasmResetNoiseTests,
                                   QasmPauliNoiseTests,
                                   QasmSnapshotStatevectorTests,
+                                  QasmSnapshotDensityMatrixTests,
                                   QasmSnapshotStabilizerTests):
     """QasmSimulator stabilizer method tests."""