+++ CHANGELOG.md
@@ -23,9 +23,11 @@ Added
 - Added tests for the Fredkin gate (#357)
 - Added tests for the cu1 gate (#360)
 - Added tests for statevector and stabilizer snapshots (\#355)
+- Added tests for density matrix snapshot (\#374)
 
 Changed
 -------
+- Change signature of density matrix snapshot extension (\#374)
 - Stabilizer snapshot returns stabilizer instead of full Clifford table (\#355)
 - Signature of SnapshotStatevector and SnapshotStabilizer (\#355)
 - Changed all names from tensor_network_state to matrix_product_state (\#356)